+++ NEWS.md
@@ -1,5 +1,9 @@
 ## HLAtools (development version)
 
+### Version 0.9.15.9000 June 21, 2024
+
+- Added addCodonLine() function to facilitate fixing nucleotide alignments that are missing "AA codon" lines.
+
 ### Version 0.9.14.9000 June 10 to June 20, 2024
 
 - Updated buildAlignments() to build DPA2 and DPB2 cDNA alignments in releases prior to 3.54.0.
@@ -20,6 +24,7 @@
 - Updated documentation for updateAll().
 - Updated repoVersion() for release 3.00.0.
 - Updated buildAlignments() to build HLA-DPA and -DPB cDA alignments in release 3.00.0.
+- Updated buildAlignments() to build TAP1 and TAP2 cDNA alignments in version 3.00.0.
 
 ### Version 0.9.13.9000 June 4, 2024